ancestral: Use parse_genes_argument
Previously, the help text for --genes falsely claimed that it could
accept a file. Now, it can truly claim that.

Move the file parsing functions to utils.
